bbf: add support for config of provider bridge residential domain scenario
This is l3 QinQ basically, the support is added for the provider bridge
residential domain scenario, details as mentioned in tr181 issue 2,
amendment 13 as follows:

In the Residential Domain scenario untagged traffic is routed from the Ethernet and SSIDa
interfaces and tagged with a customer VLAN tag (C-TAG) of VLANa and then double tagged
with a Service Provider VLAN tag (S-TAG) of VLANx. This requires the use of:
• 802.1d Bridge instance: This object bridges the residential domain traffic to the Router.
• Layer3 Forwarding Rule: This object ensures that traffic between the Bridge and
VLANTermination objects is forwarded to the correct interface. The Rule utilizes the IP
Interfaces of the Bridge (IP Intf: Bridge_a) and Residential Domain (IP Intf: Residential)
• VLANTermination object (C-TAG): The C-TAG is applied and removed for traffic
egress and ingress to the IP Intf: Residential interface.
• VLANTermination object (S-TAG): The S-TAG is applied and removed for traffic from
and to the C-VLAN termination object.

Both config and end to end traffic after setup have been tested.

Note: partially resolves ticket #3101