Ethernet.VLANTermination: set LowerLayers has no effect on the IP.Interface section related to it
if we have an IP.Interface.{i}. links to -> Ethernet.VLANTermination.{i}. which link to -> Ethernet.Link.1. and we need to update Ethernet.VLANTermination.{i}.LowerLayers to Ethernet.Link.2.

the issue is IP.Interface.{i}. was not updated when setting Ethernet.VLANTermination.{i}.LowerLayers to Ethernet.Link.2.

So this commit fixes this issue